requirements.txt
Django==3.1.6
djangorestframework==3.12.2
Markdown==3.3.3
mysqlclient==2.0.3
PyJWT==2.0.1